refactor: 연관관계 제거

+++ src/main/java/com/WalkiePaw/domain/board/repository/BoardLikeRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface BoardLikeRepository extends JpaRepository<BoardLike, Integer> {
-    BoardLike findByMemberIdAndBoardId(Integer memberId, Integer boardId);
+public interface BoardLikeRepository extends JpaRepository<BoardLike, Long> {
+    BoardLike findByMemberIdAndBoardId(Long memberId, Long boardId);
 
     @Query("select bl.board.id, count(bl) from BoardLike bl group by bl.board.id")
     List<Integer[]> countAllBoardLike();
