+++ src/main/java/com/WalkiePaw/domain/chatroom/repository/ChatroomRepositoryOverride.java
@@ -1,22 +1,20 @@
 package com.WalkiePaw.domain.chatroom.repository;
 
-import com.WalkiePaw.domain.board.entity.Board;
 import com.WalkiePaw.domain.chatroom.entity.Chatroom;
-import com.WalkiePaw.presentation.domain.chatroom.dto.ChatroomListResponse;
-import com.WalkiePaw.presentation.domain.chatroom.dto.TransactionResponse;
+import com.WalkiePaw.presentation.domain.chatroom.response.ChatroomListResponse;
+import com.WalkiePaw.presentation.domain.chatroom.response.TransactionResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.data.jpa.repository.EntityGraph;
 
 import java.util.Optional;
 
 public interface ChatroomRepositoryOverride {
-    Slice<ChatroomListResponse> findByMemberId(Integer memberId, Pageable pageable);
+    Slice<ChatroomListResponse> findByMemberId(Long memberId, Pageable pageable);
 
-    Page<TransactionResponse> findTransaction(Integer memberId, Pageable pageable);
+    Page<TransactionResponse> findTransaction(Long memberId, Pageable pageable);
 
-    Optional<Chatroom> findByMemberIdAndBoardId(Integer MemberId, Integer boardId);
+    Optional<Chatroom> findByMemberIdAndBoardId(Long MemberId, Long boardId);
 
-    Optional<Chatroom> findByWriterIdAndBoardId(Integer writerId, Integer boardId);
+    Optional<Chatroom> findByWriterIdAndBoardId(Long writerId, Long boardId);
 }
