+++ src/main/java/com/WalkiePaw/domain/member/Repository/MemberRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.Optional;
 
 @Profile("spring-data-jpa")
-public interface MemberRepository extends JpaRepository<Member, Integer>, MemberRepositoryOverride {
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryOverride {
 
     @Query("select m from Member m where m.email = :email")
     Optional<Member> findByEmail(@Param("email") String email);
