+++ src/main/java/com/WalkiePaw/presentation/domain/chat/ChatController.java
@@ -1,15 +1,10 @@
 package com.WalkiePaw.presentation.domain.chat;
 
 import com.WalkiePaw.domain.chat.service.ChatService;
-import com.WalkiePaw.presentation.domain.chat.dto.ChatAddRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import com.WalkiePaw.presentation.domain.chat.dto.ChatMsgListResponse;
+import com.WalkiePaw.presentation.domain.chat.response.ChatMsgListResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.web.bind.annotation.*;
 
@@ -25,13 +20,13 @@ public class ChatController {
     private final SimpMessagingTemplate webSocket;
 
     @GetMapping
-    public ResponseEntity<List<ChatMsgListResponse>> getChatListByChatroomId(final @RequestParam Integer chatroomId) {
+    public ResponseEntity<List<ChatMsgListResponse>> getChatListByChatroomId(final @RequestParam Long chatroomId) {
         List<ChatMsgListResponse> chats = chatService.findChatsByChatroomId(chatroomId);
         return ResponseEntity.ok(chats);
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<Void> chatMsgUpdateIsRead(final @PathVariable("id") Integer chatroomId) {
+    public ResponseEntity<Void> chatMsgUpdateIsRead(final @PathVariable("id") Long chatroomId) {
         chatService.bulkUpdateIsRead(chatroomId);
         return ResponseEntity.noContent().build();
     }
